Framework
In here lies the base framework for the quad squad. There isn't much to look at but believe me, getting this controller going proved to be far more difficult than I had anticipated but nevertheless, you can connect a controller, move the joystick and the program reads it. From here on, all that is left to do is map the buttons and start on making the game like we normally would.